Added database migration tooling and initialized db with original table.
