+++ .gitignore
@@ -13,6 +13,7 @@ __pycache__/
 .vs
 crawler.py
 
+
 # Created by https://www.gitignore.io/api/linux,macos,windows
 # Edit at https://www.gitignore.io/?templates=linux,macos,windows
