+++ src/scrapers/models/__init__.py
@@ -1,0 +1,2 @@
+from .database import Db
+from .initiatives import Initiative
